Consolidate release notes components and flatten getDeactivatedCategory()
