+++ src/apiDefs/schema.ts
@@ -49,7 +49,6 @@ export interface IApiDescription {
 export interface BaseAPICategory {
   readonly apis: IApiDescription[];
   readonly properName: string;
-  readonly buttonText: string;
   readonly name: string;
 }
